+++ ionoscloud/data_source_k8s_cluster_test.go
@@ -22,7 +22,7 @@ func TestAccDataSourceK8sCluster_matchId(t *testing.T) {
 				Config: testAccDataSourceProfitBricksK8sClusterMatchId,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "name", "test_cluster"),
-					resource.TestCheckResourceAttr("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "k8s_version", "1.20.6"),
+					resource.TestCheckResourceAttr("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "k8s_version", "1.20.8"),
 					resource.TestCheckResourceAttrSet("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "kube_config"),
 				),
 			},
@@ -44,7 +44,7 @@ func TestAccDataSourceK8sCluster_matchName(t *testing.T) {
 				Config: testAccDataSourceProfitBricksK8sClusterMatchName,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "name", "test_cluster"),
-					resource.TestCheckResourceAttr("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "k8s_version", "1.20.6"),
+					resource.TestCheckResourceAttr("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "k8s_version", "1.20.8"),
 					resource.TestCheckResourceAttrSet("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "kube_config"),
 					resource.TestCheckResourceAttrSet("data.ionoscloud_k8s_cluster.test_ds_k8s_cluster", "id"),
 				),
@@ -57,14 +57,14 @@ func TestAccDataSourceK8sCluster_matchName(t *testing.T) {
 const testAccDataSourceProfitBricksK8sClusterCreateResources = `
 resource "ionoscloud_k8s_cluster" "test_ds_k8s_cluster" {
   name         = "test_cluster"
-  k8s_version  = "1.20.6"
+  k8s_version  = "1.20.8"
 }
 `
 
 const testAccDataSourceProfitBricksK8sClusterMatchId = `
 resource "ionoscloud_k8s_cluster" "test_ds_k8s_cluster" {
   name         = "test_cluster"
-  k8s_version  = "1.20.6"
+  k8s_version  = "1.20.8"
 }
 
 data "ionoscloud_k8s_cluster" "test_ds_k8s_cluster" {
@@ -75,7 +75,7 @@ data "ionoscloud_k8s_cluster" "test_ds_k8s_cluster" {
 const testAccDataSourceProfitBricksK8sClusterMatchName = `
 resource "ionoscloud_k8s_cluster" "test_ds_k8s_cluster" {
   name         = "test_cluster"
-  k8s_version  = "1.20.6"
+  k8s_version  = "1.20.8"
 }
 
 data "ionoscloud_k8s_cluster" "test_ds_k8s_cluster" {