+++ ionoscloud/resource_datacenter.go
@@ -29,7 +29,6 @@ func resourceDatacenter() *schema.Resource {
 				Required:     true,
 				ValidateFunc: validation.All(validation.StringIsNotWhiteSpace),
 			},
-
 			"location": {
 				Type:         schema.TypeString,
 				Required:     true,